refactor(rarity-section): ♻️ remove cardStack.tsx and inline card stacking logic
