+++ apps/nextjs/src/components/pages/user/collection/index.ts
@@ -1,4 +1,3 @@
-export * from './card-stack';
 export * from './rarity-section';
 export * from './user-card-rarities';
 export * from './user-profile-image';
