+++ apps/nextjs/src/components/pages/user/collection/rarity-section.tsx
@@ -1,13 +1,6 @@
 import type { Card } from '~/common';
 import { cn } from '~/lib/utils';
-import { CardStack } from './card-stack';
-import { groupSingleCardById } from './utils/cardUtils';
-
-interface CardCountMap {
-  card: Card;
-  quantity: number;
-  isFoil: boolean;
-}
+import type { CardCountMap, CardGroup } from './utils/cardUtils';
 
 interface RaritySectionProps {
   /**
@@ -60,6 +53,62 @@ export const RaritySection = ({
 
   const cardsToDisplay = showUnknown ? cardsInRarity : ownedCardsInRarity;
 
+  const groupSingleCardById = (cardCountMap: CardCountMap): CardGroup => {
+    const { card, quantity, isFoil } = cardCountMap;
+    const cardGroup: CardGroup = { foil: null, nonFoil: null };
+
+    if (isFoil) {
+      cardGroup.foil = { ...card, quantity };
+    } else {
+      cardGroup.nonFoil = { ...card, quantity };
+    }
+
+    return cardGroup;
+  };
+
+  const renderCardStack = (card: Card) => {
+    const ownedCard = ownedCardMap.get(card.id);
+    const isOwned = !!ownedCard;
+    const cardGroup = ownedCard ? groupSingleCardById(ownedCard) : { foil: null, nonFoil: null };
+
+    const mainCard = cardGroup.foil ?? cardGroup.nonFoil ?? card;
+    const mainCardCount = cardGroup.foil?.quantity ?? cardGroup.nonFoil?.quantity ?? 0;
+
+    const limitedDuplicates = mainCardCount > 1 && isOwned ? Array(Math.min(mainCardCount - 1, 4)).fill(null) : [];
+
+    const getCardStackClasses = (stackIndex: number) =>
+      cn('absolute w-full h-full top-0 left-0 bg-cover transition-transform ease-elastic duration-500', {
+        'group-hover:-translate-y-4 group-hover:rotate-[-4deg] group-hover:-translate-x-2 z-[10]': stackIndex === 3,
+        'group-hover:translate-y-4 group-hover:rotate-[7deg] z-[15]': stackIndex === 2,
+        'group-hover:-translate-x-4 group-hover:rotate-[-7deg] z-[13]': stackIndex === 1,
+        'group-hover:translate-x-6 group-hover:rotate-[5deg] z-[16]': stackIndex === 0,
+      });
+
+    const cardContainerClasses = cn('relative', {
+      'hover:scale-105 transition-transform ease-elastic group duration-500 hover:z-50': isOwned,
+      grayscale: !isOwned,
+    });
+
+    return (
+      <div key={card.id} className={cardContainerClasses}>
+        {/* Render stacked duplicates behind the main card */}
+        {isOwned &&
+          limitedDuplicates.map((_, stackIndex) => (
+            <div key={stackIndex} className={getCardStackClasses(stackIndex)}>
+              <img src={mainCard.image} alt={`Duplicate ${stackIndex + 1}`} className="w-60 h-82 card-image z-10" />
+            </div>
+          ))}
+
+        {/* Main Card */}
+        <img
+          src={mainCard.image}
+          alt={mainCard.name}
+          className={cn('w-60 h-82 card-image relative z-20', { 'opacity-50': !isOwned })}
+        />
+      </div>
+    );
+  };
+
   return (
     <div key={rarity} id={rarity.toLowerCase()} className={classes.sectionContainer}>
       <h2 className={classes.title}>
@@ -69,14 +118,7 @@ export const RaritySection = ({
         </span>
       </h2>
 
-      <div className={classes.cardGrid}>
-        {cardsToDisplay.map((card) => {
-          const ownedCard = ownedCardMap.get(card.id);
-          const cardGroup = ownedCard ? groupSingleCardById(ownedCard) : { foil: null, nonFoil: null };
-
-          return <CardStack key={card.id} cardGroup={cardGroup} card={card} isOwned={!!ownedCard} />;
-        })}
-      </div>
+      <div className={classes.cardGrid}>{cardsToDisplay.map((card) => renderCardStack(card))}</div>
     </div>
   );
 };
